requirements.txt
fastapi==0.115.5
uvicorn==0.31.1
beanie==1.27.0
motor==3.6.0
pyton-dotenv==1.0.1